+++ library/src/main/kotlin/eu/tib/orkg/prototype/export/rdf/api/ExportRDFUseCase.kt
@@ -7,6 +7,7 @@ import org.eclipse.rdf4j.model.Model
 
 interface ExportRDFUseCase {
     fun dumpToNTriple(writer: Writer)
+    fun dumpToNTriple(path: String?)
     fun rdfModelForClass(id: ThingId): Optional<Model>
     fun rdfModelForPredicate(id: ThingId): Optional<Model>
     fun rdfModelForResource(id: ThingId): Optional<Model>