+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- The RDF export does not use the atomic move operation anymore. (See: [!483](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/483))
 ### Changed
 - The benchmark endpoints are now paged. (Closes: [#461](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/461))